Updated test folder in backend to have correct naming
